refactor(stats): move capture control-plane SPI to core stats
